Refactor + simplify
