+++ packages/front-end/pages/features/[fid].tsx
@@ -1,20 +1,6 @@
 import { useRouter } from "next/router";
 import { useEffect, useState, useMemo } from "react";
-import { FeatureInterface, FeatureRule } from "shared/types/feature";
-import { FeatureCodeRefsInterface } from "shared/types/code-refs";
-import { FeatureRevisionInterface } from "shared/types/feature-revision";
-import { ExperimentInterfaceStringDates } from "shared/types/experiment";
-import {
-  filterEnvironmentsByFeature,
-  getDependentExperiments,
-  getDependentFeatures,
-  mergeRevision,
-} from "shared/util";
-import {
-  SafeRolloutInterface,
-  HoldoutInterface,
-  MinimalFeatureRevisionInterface,
-} from "shared/validators";
+import { getDependentExperiments, getDependentFeatures } from "shared/util";
 import { FeatureEvalDiagnosticsQueryResponseRows } from "shared/types/integrations";
 import LoadingOverlay from "@/components/LoadingOverlay";
 import PageHead from "@/components/Layout/PageHead";
@@ -23,7 +9,7 @@ import { useLocalStorage } from "@/hooks/useLocalStorage";
 import FeaturesOverview from "@/components/Features/FeaturesOverview";
 import FeaturesStats from "@/components/Features/FeaturesStats";
 import useOrgSettings from "@/hooks/useOrgSettings";
-import { useEnvironments, useFeaturesList } from "@/services/features";
+import { useFeaturesList } from "@/services/features";
 import { FeatureUsageProvider } from "@/components/Features/FeatureUsageGraph";
 import FeatureTest from "@/components/Features/FeatureTest";
 import { useAuth } from "@/services/auth";
@@ -31,27 +17,11 @@ import EditTagsForm from "@/components/Tags/EditTagsForm";
 import EditFeatureInfoModal from "@/components/Features/EditFeatureInfoModal";
 import { useExperiments } from "@/hooks/useExperiments";
 import FeatureDiagnostics from "@/components/Features/FeatureDiagnostics";
-import useApi from "@/hooks/useApi";
+import { useFeaturePageData } from "@/hooks/useFeaturePageData";
+import Callout from "@/ui/Callout";
 
 const featureTabs = ["overview", "stats", "test", "diagnostics"] as const;
 export type FeatureTab = (typeof featureTabs)[number];
-
-type FeaturePageResponse = {
-  feature: FeatureInterface | null;
-  revisionList: MinimalFeatureRevisionInterface[];
-  revisions: FeatureRevisionInterface[];
-  experiments: ExperimentInterfaceStringDates[];
-  safeRollouts: SafeRolloutInterface[];
-  codeRefs: FeatureCodeRefsInterface[];
-  holdout: HoldoutInterface | undefined;
-};
-
-function parseVersion(value: string | string[] | undefined): number | null {
-  const v = Array.isArray(value) ? value[0] : value;
-  if (!v) return null;
-  const parsed = parseInt(v, 10);
-  return Number.isNaN(parsed) ? null : parsed;
-}
 
 export default function FeaturePage() {
   const router = useRouter();
@@ -60,162 +30,30 @@ export default function FeaturePage() {
   const [editProjectModal, setEditProjectModal] = useState(false);
   const [editTagsModal, setEditTagsModal] = useState(false);
   const [editFeatureInfoModal, setEditFeatureInfoModal] = useState(false);
-  const [version, setVersion] = useState<number | null>(null);
   const [diagnosticsResults, setDiagnosticsResults] = useState<Array<
     FeatureEvalDiagnosticsQueryResponseRows[number] & { id: string }
   > | null>(null);
-
-  // To ensure that when we navigate between versions we don't refetch them if it is not needed
-  const [cachedRevisionsByVersion, setCachedRevisionsByVersion] = useState<
-    Record<number, FeatureRevisionInterface>
-  >({});
-  const [cachedExperimentsById, setCachedExperimentsById] = useState<
-    Record<string, ExperimentInterfaceStringDates>
-  >({});
-  const [cachedSafeRolloutsById, setCachedSafeRolloutsById] = useState<
-    Record<string, SafeRolloutInterface>
-  >({});
+  // Clean state when feature id changes
+  useEffect(() => {
+    setDiagnosticsResults(null);
+  }, [fid]);
 
   const { apiCall } = useAuth();
   const { experiments: allExperiments } = useExperiments();
 
-  const forcedVersionFromQuery = useMemo(
-    () => parseVersion(router.query.v),
-    [router.query.v],
-  );
-  const selectedVersion = version ?? forcedVersionFromQuery;
-
   const {
-    data: baseData,
-    error: baseError,
-    mutate: mutateBase,
-    isValidating: isValidatingBase,
-  } = useApi<FeaturePageResponse>(fid ? `/feature/${fid}` : "", {
-    shouldRun: () => !!fid,
-  });
+    data,
+    error,
+    isValidating,
+    refreshData,
+    feature,
+    baseFeature,
+    revision,
+    environments,
+    version,
+    setVersion,
+  } = useFeaturePageData(fid, router.query.v);
 
-  const shouldFetchSelectedVersion =
-    fid !== undefined &&
-    selectedVersion !== null &&
-    !(
-      (baseData?.revisions ?? []).some((r) => r.version === selectedVersion) ||
-      !!cachedRevisionsByVersion[selectedVersion]
-    );
-
-  const {
-    data: selectedVersionData,
-    error: selectedVersionError,
-    mutate: mutateSelectedVersion,
-    isValidating: isValidatingSelectedVersion,
-  } = useApi<FeaturePageResponse>(
-    fid && selectedVersion ? `/feature/${fid}?v=${selectedVersion}` : "",
-    {
-      shouldRun: () => shouldFetchSelectedVersion,
-    },
-  );
-
-  // Ensure we reset everything if feature id changes
-  useEffect(() => {
-    if (!fid) return;
-    setVersion(null);
-    setDiagnosticsResults(null);
-    setCachedRevisionsByVersion({});
-    setCachedExperimentsById({});
-    setCachedSafeRolloutsById({});
-  }, [fid]);
-
-  const refreshData = async () => {
-    await mutateBase();
-    if (shouldFetchSelectedVersion) {
-      await mutateSelectedVersion();
-    }
-  };
-
-  useEffect(() => {
-    if (!baseData || !baseData.feature || baseData.feature.id !== fid) {
-      return;
-    }
-
-    setCachedRevisionsByVersion((prev) => {
-      const next = { ...prev };
-      baseData.revisions.forEach((r) => {
-        next[r.version] = r;
-      });
-      return next;
-    });
-
-    setCachedExperimentsById((prev) => {
-      const next = { ...prev };
-      baseData.experiments.forEach((e) => {
-        next[e.id] = e;
-      });
-      return next;
-    });
-
-    setCachedSafeRolloutsById((prev) => {
-      const next = { ...prev };
-      baseData.safeRollouts.forEach((sr) => {
-        next[sr.id] = sr;
-      });
-      return next;
-    });
-  }, [baseData, fid]);
-
-  useEffect(() => {
-    if (
-      !selectedVersionData ||
-      !selectedVersionData.feature ||
-      selectedVersionData.feature.id !== fid
-    ) {
-      return;
-    }
-
-    setCachedRevisionsByVersion((prev) => {
-      const next = { ...prev };
-      selectedVersionData.revisions.forEach((r) => {
-        next[r.version] = r;
-      });
-      return next;
-    });
-    setCachedExperimentsById((prev) => {
-      const next = { ...prev };
-      selectedVersionData.experiments.forEach((e) => {
-        next[e.id] = e;
-      });
-      return next;
-    });
-    setCachedSafeRolloutsById((prev) => {
-      const next = { ...prev };
-      selectedVersionData.safeRollouts.forEach((sr) => {
-        next[sr.id] = sr;
-      });
-      return next;
-    });
-  }, [selectedVersionData, fid]);
-
-  const data = useMemo<FeaturePageResponse | undefined>(() => {
-    const source = baseData ?? selectedVersionData;
-    if (!source) return undefined;
-
-    return {
-      ...source,
-      revisions: Object.values(cachedRevisionsByVersion),
-      experiments: Object.values(cachedExperimentsById),
-      safeRollouts: Object.values(cachedSafeRolloutsById),
-    };
-  }, [
-    baseData,
-    selectedVersionData,
-    cachedRevisionsByVersion,
-    cachedExperimentsById,
-    cachedSafeRolloutsById,
-  ]);
-  const error = selectedVersionError ?? baseError;
-  const isValidating = isValidatingBase || isValidatingSelectedVersion;
-
-  const baseFeature = data?.feature;
-  const baseFeatureVersion = baseFeature?.version;
-  const revisions = data?.revisions;
   const experiments = data?.experiments;
   const safeRollouts = data?.safeRollouts;
   const holdout = data?.holdout;
@@ -225,7 +63,6 @@ export default function FeaturePage() {
     project: baseFeature?.project,
     skipFetch: !baseFeature,
   });
-  const allEnvironments = useEnvironments();
 
   const [tab, setTab] = useLocalStorage<FeatureTab>(
     `tabbedPageTab__${fid}`,
@@ -255,90 +92,7 @@ export default function FeaturePage() {
     return () => window.removeEventListener("hashchange", handler, false);
   }, [setTab]);
 
-  // Set the initial selected version once data is available.
-  useEffect(() => {
-    if (!baseFeatureVersion || version !== null) return;
-
-    if (forcedVersionFromQuery) {
-      if (
-        revisions &&
-        revisions.some((r) => r.version === forcedVersionFromQuery)
-      ) {
-        setVersion(forcedVersionFromQuery);
-      }
-      return;
-    }
-
-    // If there's an active draft, show that by default, otherwise show the live version
-    const draft =
-      revisions &&
-      revisions.find(
-        (r) =>
-          r.status === "draft" ||
-          r.status === "approved" ||
-          r.status === "changes-requested" ||
-          r.status === "pending-review",
-      );
-    setVersion(draft ? draft.version : baseFeatureVersion);
-  }, [revisions, version, forcedVersionFromQuery, baseFeatureVersion]);
-
-  const environments = useMemo(
-    () =>
-      baseFeature
-        ? filterEnvironmentsByFeature(allEnvironments, baseFeature)
-        : [],
-    [allEnvironments, baseFeature],
-  );
   const envs = environments.map((e) => e.id);
-
-  const revision = useMemo<FeatureRevisionInterface | null>(() => {
-    if (!baseFeature) return null;
-
-    const currentVersion =
-      version ?? forcedVersionFromQuery ?? baseFeature.version ?? null;
-
-    if (!currentVersion) return null;
-
-    const match =
-      revisions && revisions.find((r) => r.version === currentVersion);
-    if (match) {
-      return match;
-    }
-
-    // If we can't find the revision, create a dummy revision just so the page can render
-    // This is for old features that don't have any revision history saved
-    const rules: Record<string, FeatureRule[]> = {};
-    environments.forEach((env) => {
-      rules[env.id] = baseFeature.environmentSettings?.[env.id]?.rules || [];
-    });
-    return {
-      baseVersion: baseFeature.version,
-      comment: "",
-      createdBy: null,
-      dateCreated: baseFeature.dateCreated,
-      datePublished: baseFeature.dateCreated,
-      dateUpdated: baseFeature.dateUpdated,
-      defaultValue: baseFeature.defaultValue,
-      featureId: baseFeature.id,
-      organization: baseFeature.organization,
-      publishedBy: null,
-      rules: rules,
-      status: "published",
-      version: baseFeature.version,
-      prerequisites: baseFeature.prerequisites || [],
-    };
-  }, [revisions, version, forcedVersionFromQuery, environments, baseFeature]);
-
-  const feature = useMemo(() => {
-    if (!revision || !baseFeature) return null;
-    return revision.version !== baseFeature.version
-      ? mergeRevision(
-          baseFeature,
-          revision,
-          environments.map((e) => e.id),
-        )
-      : baseFeature;
-  }, [baseFeature, revision, environments]);
 
   // note: project-scoped dependents by default
   const dependentFeatures = useMemo(() => {
@@ -354,11 +108,7 @@ export default function FeaturePage() {
   const dependents = dependentFeatures.length + dependentExperiments.length;
 
   if (error) {
-    return (
-      <div className="alert alert-danger">
-        An error occurred: {error.message}
-      </div>
-    );
+    return <Callout status="error">An error occurred: {error.message}</Callout>;
   }
 
   if (!data || !feature || !revision || !baseFeature) {
